use hlcoop version of PrecacheSound instead
the engine version was causing crashes because hlcoop was unaware of the sound being precached and couldn't detect the overflow. For hl and Sven Co-op the sound is always precached.

Anyone building this for HL/Sven Co-op should remove the LOAD_GAME_API def in CMakeLists.txt unless you don't mind a metamod error every time the plugin loads.